# docs/swagger.yaml
definitions:
  controller.CustomerInput:
    properties:
      address:
        type: string
      name:
        type: string
    type: object
...
  controller.SellerInput:
    properties:
      address:
        type: string
      name:
        type: string
    type: object
  controller.TransactionInput:
    properties:
      description:
        type: string
      productID:
        type: integer
      quantity:
        type: integer
      status:
        type: string
    type: object
...
  models.Seller:
    properties:
      address:
        type: string
      id:
        type: integer
      name:
        type: string
      userID:
        type: integer
    type: object
  models.Transaction:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      productID:
        type: integer
      quantity:
        type: integer
      sellerID:
        type: integer
      string:
        type: string
      total_price:
        type: integer
      updated_at:
        type: string
      userID:
        type: integer
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
paths:
...
  /seller:
    get:
      description: Get a list of Seller
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Seller'
            type: array
      summary: Get all Seller.
      tags:
      - Seller
    post:
      description: Creating a new Seller.
      parameters:
      - description: the body to create a new Seller
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.SellerInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Seller'
      security:
      - BearerToken: []
      summary: Register user as a Seller (Guest Only).
      tags:
      - Seller
  /seller/{id}:
    delete:
      description: Delete a Seller by id.
      parameters:
      - description: Seller id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Seller (Referred Seller Only).
      tags:
      - Seller
    get:
      description: Get an Seller by id.
      parameters:
      - description: Seller id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Seller'
      summary: Get Seller by Id.
      tags:
      - Seller
    patch:
      description: Update seller by id.
      parameters:
      - description: Seller id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update seller
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.CustomerInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Seller'
      security:
      - BearerToken: []
      summary: Update Seller (Referred Seller Only).
      tags:
      - Seller
  /transaction:
    get:
      description: Get a list of Transaction.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Transaction'
            type: array
      security:
      - BearerToken: []
      summary: Get all Transaction (Admin Only).
      tags:
      - Transaction
    post:
      description: Creating a new Transaction.
      parameters:
      - description: the body to create a new Product
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.TransactionInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
      security:
      - BearerToken: []
      summary: Create a new Transaction (Customer Only).
      tags:
      - Transaction
  /transaction/{id}:
    delete:
      description: Delete a Transaction by id.
      parameters:
      - description: Transaction id
        in: path
        name: id
        required: true
        type: string
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - BearerToken: []
      summary: Delete one Transaction (Admin Only).
      tags:
      - Transaction
    get:
      description: Get an Transaction by id.
      parameters:
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      - description: Transaction id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
      security:
      - BearerToken: []
      summary: Get Transaction By Id (Admin and Customer in Transaction Only).
      tags:
      - Transaction
    patch:
      description: Update Transaction by id.
      parameters:
      - description: Transaction id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update transaction
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controller.TransactionInput'
      - description: 'Authorization. How to input in swagger : ''Bearer <insert_your_token_here>'''
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Transaction'
      security:
      - BearerToken: []
      summary: Update Transaction (Customer and Seller in Transaction Only).
      tags:
      - Transaction
...
swagger: "2.0"